+++ packages/util-crypto/src/xxhash/asU8a.spec.ts
@@ -4,7 +4,7 @@
 import { hexToU8a } from '@polkadot/util';
 import { waitReady } from '@polkadot/wasm-crypto';
 
-import { performanceWasm } from '../test/performance';
+import { perfWasm } from '../test';
 import { xxhashAsU8a } from '.';
 
 describe('xxhashAsU8a', (): void => {
@@ -38,7 +38,7 @@ describe('xxhashAsU8a', (): void => {
     });
   });
 
-  performanceWasm('xxhashAsU8a', 8000, (input, onlyJs) =>
+  perfWasm('xxhashAsU8a', 8000, (input, onlyJs) =>
     xxhashAsU8a(input, 128, onlyJs)
   );
 });