+++ packages/util/src/u8a/toBigInt.spec.ts
@@ -1,7 +1,7 @@
 // Copyright 2017-2023 @polkadot/util authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { perf } from '../test/performance';
+import { perf } from '../test';
 import { u8aToBigInt } from '.';
 
 describe('u8aToBigInt', (): void => {