+++ packages/util/src/u8a/toHex.spec.ts
@@ -1,7 +1,7 @@
 // Copyright 2017-2023 @polkadot/util authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { perfCmp } from '../test/performance';
+import { perfCmp } from '../test';
 import { u8aToHex as u8aToHexBuffer } from './toHexBuffer';
 import { u8aToHex } from '.';