+++ packages/x-randomvalues/src/base64.ts
@@ -9,6 +9,14 @@
 // We provide an own (tiny) decoder to not have Buffer deps
 
 const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/';
+const map = new Uint8Array(256);
+
+// We use charCodeAt for access here and in the decoder loop - this is faster
+// on lookups (array + numbers) and also faster than accessing the specific
+// character via data[i] (around a 50% improvement on perf tests)
+for (let i = 0; i < chars.length; i++) {
+  map[chars.charCodeAt(i)] = i;
+}
 
 export function base64Decode (data: string): Uint8Array {
   const bytes = [];
@@ -17,7 +25,7 @@ export function base64Decode (data: string): Uint8Array {
 
   for (let i = 0; i < data.length && data[i] !== '='; i++) {
     // each character represents 6 bits
-    byte = (byte << 6) | chars.indexOf(data[i]);
+    byte = (byte << 6) | map[data.charCodeAt(i)];
 
     // each byte needs to contain 8 bits
     if ((bits += 6) >= 8) {